Import test of module execution with AutomagicIO from -core
This test will fail when uses with a datalad version that still has
AutomagicIO support on its own, but should pass for any version that
has it removed.

Given that this is just a pecularity of the test setup, and the actual
functionality works in both cases, I will not adjust versioned
dependencies, but merely skip the test.