skip certain tests on windows
This commit skips a number of tests on windows by
specifically checking for a windows-environment.
This has become necessary with the integration of
pytest in `datalad` because this repository still
uses `nose` and the pytest-skip exception is not
recognized by `nose`.

This commit can be reverted, once `datalad_deprecated`
is moved over to pytest